ai-service/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alembic==1.12.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
datasets==3.5.0
dill==0.3.8
exceptiongroup==1.2.2
faiss-cpu==1.10.0
fastapi==0.100.0
filelock==3.18.0
frozenlist==1.5.0
fsspec==2024.12.0
google-auth==2.38.0
google-genai==1.8.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.1
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
Mako==1.3.9
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.3.2
multiprocess==0.70.16
networkx==3.4.2
numpy==1.26.0
packaging==24.2
pandas==2.0.1
propcache==0.3.1
psycopg2-binary==2.9.6
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.1
pydantic_core==2.33.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.27.0
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.50.3
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.23.0
websockets==15.0.1
xxhash==3.5.0
yarl==1.18.3
